core/rawdb: fix lint nits

+++ core/rawdb/accessors_chain.go
@@ -318,8 +318,8 @@ func ReadHeaderRange(db ethdb.Reader, number uint64, count uint64) []rlp.RawValu
 		return rlpHeaders
 	}
 	// read remaining from ancients
-	max := uint64(count * 700)
-	data, err := db.AncientRange(freezerHeaderTable, uint64(i+1-count), count, max)
+	max := count * 700
+	data, err := db.AncientRange(freezerHeaderTable, i+1-count, count, max)
 	if err == nil && uint64(len(data)) == count {
 		// the data is on the order [h, h+1, .., n] -- reordering needed
 		for i := range data {
